server_side/requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
pillow==11.2.1
pip==25.0
psycopg2==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
setuptools==75.8.0
sqlparse==0.5.3
tzdata==2025.1
wheel==0.45.1
